Added EnumInfo and VariantInfo types.
These types are intended to capture the relevant parts of the AST for
the remaining operation in the proc_macro. They are generated from the
syn::DeriveInput by visiting its AST and taking references to the
relvant parts of the AST. The AST visitors also validate that the AST
has a "shape" that we can handle (and panic with error messages if it
doesn't).