Implement the code generation for the Lexer and dfa.
The code generation takes the enum_info::EnumInfo geneated from the
DeriveInput and the redfa::Dfa that will eventually be geneated from the
EnumInfo and uses that to generate a Dfa enum, to implement
luther::dfa::Dfa on the Dfa enum, and to implement luther::Lexer on the
type named in the EnumInfo using the Dfa enum.

Currently the redfa::Dfa that is used is a dummy one just for testing.
The next step will be to propery generate the redfa::Dfa from the
information in the enum_info::EnumInfo.