Added SpannedUtf8Iter and the associated reader extention.
SpannedUtf8Iter takes a fallible u8 iterator (with std::io::Error's for
its failures) and decodes utf8 encoded chars from it. It wraps the chars
in a Span to track the starting and ending locations for each char.

The reader extention (Utf8SpannedChars) provides easy access to a
SpannedUtf8Iter for every reader.