Changed the coveralls integraion on Travis CI.
Added a dependancy on the libiberty-dev package. This package appears to
be necessary to kcov (used by the cargo-travis "cargo coveralls"
command) to build in a way that supports the "--verify" option.

Removed the "--all" option from the "cargo coveralls" command. This
undoes a previous change that added it and restricts the test coverage
checking to the "luther" project. The "luther-derive" project is not
currently check for test coverage.

This changes was made for 2 reason. The first is that "cargo coveralls"
didn't work for the "luther-derive" project. It resulted in some of the
kcov invocations ending in an error. The second (and more fundamental)
is that even if we could get "cargo coveralls" to work for the
"luther-derive" project, this would test the coverage for the generated
code and not the proc-macro code. This coverage information would be
useless.