Adjusted the Travis CI build process.
Removed the apt package addon for an updated cmake package because the
build is assumed to be using the "trust" Ubuntu release which has a
verson of cmake which is more recent than the one in the kalakris-cmake
source.

Changed the "before_script" to be a "before_install" because that fits
better with the purpose of the command in that section (install the
dependancies for the build).

Changed the "script" to be an "install" and then a "script". This more
closely reflects the default travis.yml for rust which has an install of
"cargo build --verbose" and a script of "cargo test --verbose". The
install and script sections here just add the "--all" option to the
respective comments to make them apply to all packages in the workspace.