Fixed error with the visibility of the generated dfa enum.
The bug shows up if the token enum has any visibility other than
Inherited (private). In such a case the geneated code will fail to
compile with an error indicating that the private type of the dfa enum
was in a public interface. The error would show a hint of "can't leak
private type" on the "Lexer" of #[derive(Lexer)].

The fix is to make the visibility of the dfa enum match that of the
token enum. We track the visibility of the token enum in the enum_info
module and use it to set the visibility of the dfa enum in the generate
module.