Excluded defencive coding from kcov coverage.
The excluded lines and region are ones that other execution paths make
impossible to execute and, thus, impossible to test. visit_data_union()
can never be called because the compiler will intercept a proc-macro
dervie on a union before the proc-macro crate ever gets called. The line
in visit_nested_meta() that is excluded is guard for a condition that
will be caught earlier in the call chain.

It is worth leaving these two parts of the code in the codebase in case
the assumptions that currently prevent them from being executed ever
change.