Added line and region coverage exclusion.
The Travis check_test_coverage script will now recognize lines with
COV_EXCL_LINE as lines not to count for the purpose of test coverage and
regions between COV_EXCL_START and COV_EXCL_END as regions not to count
for this purpose.

This change (together with the use of these markers in the source code)
will allow the project to define the goal of "high test coverage" as
100% test coverage reported by coveralls.io together with a low number
of uses of the markers.

This is an improvement over defining "high test coverage" as (for
example) "at least 90% coverage reported by coveralls.io" because it
allows us to record our judgement about why coverage of a specific line
or region is not needed in the commit message that adds a use of the
markers.