Fixed support for running testsuite under kcov.
Added support for creating an empty kcov-out directory when that option
is given on the command line.

Also added some trace output to be better able to diagnose problems
encountered with running the testsuite.

Refactored the test runner to extract the command being run into a new
type (runner::Command) that wraps std::process::Command to allow for
cleaner logging of debug and trace output related to the execution of
each command.