Updated the project level documentation.
Added a Code of Conduct and Contributing instructions. Updated the top
level README to reference the Code of Conduct and the Contributing
instructions. Also updated the top level README to provide a better high
level description of the point of the project.

Updated the testsuite README to describe the new cargo alias for running
the testsuite.

[skip ci]