Initial implementation of a PartionMap over an Alphabet.
The Alphabet trait marks the universe of values for which a PartionMap
maps. It is implemented for u8 and char.

A PartionMap is an effecient mapping from each of the elements in an
Alphabet to some value. It permits splitting the map at a specific value
and getting the value for a specific element in the Alphabet. If the
type of the values is bool then it also allows getting the union of two
PartitionMap's. This latter is intented for the implementation of a
future PartitionSet type.