Added property test for min_value invarient.
Added the strategy functions to drive the property tests for
PartitionMap as well.

The prop_partition_map_contains_min_value() property tests runs a
series of random split() and union() operations against a random
starting PartitionMap and then checks that the PartitionMap contains an
interval [0, x) for some x. This is invarient 1 for type U = u8.