Fixed corner case uncovered by property test.
This appears to be the last bug to be uncovered by the
prop_partition_map_values_alternate property test. This property test is
designed to test the second invarient: that the values in the underlying
BTreeMap that implements the PartitionMap are non-consecutive. That is,
that no consecutive values occur in the BTreeMap.