Implemented RegexContext::class().
RegexContext::class() takes an iterator of closed ranges of elements
from the alphabet and produces a new Regex of type RegexKind::Class
that uses these ranges to itentify the subset of the alphabet that
should be matched.

The implementaion of RegexContext::class() necessitated making
RegexContext, Regex, and RegexKind generic over the alphabet (identified
by the Alphabet trait).

This commit also includes some miscellaneous cleanup to Cargo.toml and
lib.rs.