Added trait Label for DFA states labeled with a Regex.
The trait Label defines what is needed of a type to act as a label for a
state in the generation of a DFA. This is a sealed trait restricted to
types in this crate.

The rational for sealing Label is that the DFA construction algorithm
makes certain assumption about the types that are used as the labels
that would be hard to specify except by refering to Owens et al. Since
this crate is intented to fully implement the two types of lables
described in Owens et al. (that is, regular expressions and regular
vectors) there is little to be gained by making Label an unsafe trait
with a long description of the assumptions that must be satisfied by an
implementation.

This commit also implements Label for Regex using the rules described in
section 3.1 and 4.2 of Owens et al.

The commit removes some of the Debug constraints on the implementation
of some other methods of other types to simplify Label and its
implementations.