Added derivative_classes() to the StateLabel trait.
derivative_classes() returns the DerivativeClasses for a value of the
type for which StateLabel is implemented (Regex or in future regular
vectors). The DerivativeClasses are a partition of the Alphabet into
subsets that have the characteristic that all the members of the subset
have equivalant derivatives, and hence have the same transition to the
next state in the generated deterministic finite automaton.

Since the subsets in DerivativeClasses intentify different
transitions out of the state labeled with the StateLabel, these subsets
are themselves labeled with a TransitionLabel.

The derivative_classes() method is the C(r) function from section 4.2 of
Owen et al. In the case of Regex, it is using structural recursion to
build an approximation of the true derivitive classes.

The PartitionMap::from_merge() constructor is the ∧ function from
section 4.2 of Owen et al.

DerivativeClasses are implemented as a PartitionMap from the Alphabet to
a TransitionLabel. In the general case, TransitionLabel's will repeat in
the PartitionMap because more than one half-open range will be part of
the same subset of the Alphabet.