Outline Dfa types and implement dfa::State.
A Dfa is a collection of State's. The State type is implemented with
this commit, but the Dfa type itself remains a skeleton that is
yet-to-be-implemented.

This commit also adds a Hash implementation to many types so that
StateLabel can inherit from Hash, which will be needed for the dfa
construction algorithm.