Implemented Dfa methods and functions.
The Dfa::new() function is the central method of this whole crate. It is
the DFA construction algorithm using RE derivates and character classes
from Figure 3 of Owens et al.

The remaining memthods on Dfa are accessors for the DFA generated by
Dfa::new().

The commit also adds Eq as a super trait for StateLabel and implements
Eq for a number of types to the Regex can continue to implement
StateLabel. This is necessary so we can use a StateLabel as a key for a
HashMap.