Added DfaContext and code to parse regex's.
DfaContext has methods to create a Dfa from a textual representation of
a single regular expression and from an iterator of textual
representations of regular expressions. The former creates a Dfa over a
Regex. The latter creates a Dfa over a RegexVec. Both use char as the
alphabet for the Dfa.

Starting from a textual representation of regular expressions requires
parsing those representation. Such parsing is done by the regex_syntax
crate, though some regular expressions parsed by that crate result in
errors because we don't (yet) have a way of tranlating them into a
Regex.

The parsing code added with this commite can result in an error if the
textual representation of the regular expression is not well formed or
if it respresents some kind of regular expression that we can't (yet)
translate into a Regex.