Reorganize the types at the root of the crate.
The new organization highlight the usual entry points for a new user of
the library while relegating the exposed parts of the implementation to
modules.

This commit also changes some of the doc comments that referenced `Dfa`
to instead reference "dfa" (that is, the concept of a dfa rather than
its instantiation in the `Dfa` type).