Changed Regex to special case empty Regex.
The RegexContext::empty() constructor now returns a Regex::empty() which
internally stores a special Empty RegexType (an emum that is internal to
the regex::handle module). This avoids allocating an entry in the Arena
contained in the RegexContext.

This change resulted in an almost 50% speed up on the empty_regex
microbenchmark, though it does not appear to have made a speed
difference on the basis_benchmark macro benchmarks.