Changed representation of an empty regex::Class.
The old representation used an Rc counted PartitionSet that has a single
empty partition. The new representation uses Option::None (and
Option::Some() for the non-empty case).

An empty Class is the null set and forms a part of the null regular
expression that matches nothing. The null regular expression is used
frequently in the dfa construction algorithm (especially in the label
module).

This change resulted in a 73% speed improvement in the null_regex
microbenchmark and a 35-50% increase in the basic_benchmark macro
benchmarks.