Simplifed PartitionSet constructor.
The new() constructor had the full power of a RangeArgument to describe
the set being constructed. The only non-test use of this, though was for
the range "..". This change simplifies the constructor to the one case
we need. This also changed the name to "full_singleton" to call out the
kind of PartitionSet it constructs.