Delete PartionMap methods only used in tests.
The PartitionMap::get() and PartitionMap::split() methods were only used
in tests so they (and their direct tests) have been deleted. This
simplifies the interface for PartitionMap.

The PartitionMap::get() method was used in test of other parts of
PartitionMap and in tests of PartitionSet so it is replaced by a testing
only helper function map_get() which has access to the implementation
details of PartionMap.