Extracted iterator from PartionMap::from_merge().
The lazy iterator is now in the partition::merge_iter() function while
the eager PartionMap::from_merge() function uses the
PartionMap::from_lower_bound_iter() constructor with the iterator
returned by merge_iter().

This commit extracts the lazy iterator from the eager constructor and is
a step towards relying on laziness as a performance enchancement for
PartitionMap construction.