Renamed PartitionMap and PartitionSet.
The new names (Partition and Set respectivly) better reflect the purpose
of these two types. A Set (formerly PartitionSet) is not concerned with
a partition of anything, but rather is just an implementaion of the
abtract type "set". A Partition (formerly PartitionMap) is precisely an
implementation of a abstract "partition" of some universe U into subsets
that are identified by having the same value `V`. It doesn't really map
`U` to `V`, but rather uses values of `V` to identify the subsets of
`U`.